Charts are sorted by correct index again
added _alt array to hold arrays of a chart and its index (because any other solution I tried failed, storing the index in an array with the chart was my only working solution).

Made sorting method for [[thing,index],[thing2,index]] to sort by index in the array and return it to a variable.

menu reloads when theres an error loading charts from the server for a more seamless transition into the actual game.

Fixed small issue where not all maps would load and you'd have to enter something into the search bar to fix it